# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RESIT_Immo API",
        "description": "API REST pour la gestion locative — propriétaires, biens, locataires et contrats.",
        "contact": {
            "name": "Support RESIT_Immo",
            "email": "support@resit-immo.fr"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
...
        "/api/locataires": {
            "get": {
                "tags": [
                    "Locataires"
                ],
                "summary": "Lister tous les locataires",
                "operationId": "3736e06fb5ee490cecdec8d27dcbe569",
                "responses": {
                    "200": {
                        "description": "Liste de tous les locataires",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Locataire"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Locataires"
                ],
                "summary": "Créer un locataire",
                "operationId": "1a2bf3b9521be15a43a650b5dbab61bd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Durand"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Alice"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "alice@example.com",
                                        "nullable": true
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "07 98 76 54 32",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Locataire créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Locataire créé avec succès."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Locataire"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/locataires/{id}": {
            "get": {
                "tags": [
                    "Locataires"
                ],
                "summary": "Afficher un locataire avec ses contrats",
                "operationId": "e162ae49c53ee8fd8b8403ab987532de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du locataire",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail du locataire avec ses contrats",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Locataire"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Locataire introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Locataires"
                ],
                "summary": "Mettre à jour un locataire (remplacement complet)",
                "operationId": "6f4007a0b66cb4bf2697b29cfb094281",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du locataire",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Durand"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Alice"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Locataire mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Locataire mis à jour avec succès."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Locataire"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Locataire introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Locataires"
                ],
                "summary": "Supprimer un locataire",
                "operationId": "8b3cd0a748dfa80198b4d3f075caf41f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du locataire",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Locataire supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Locataire supprimé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Locataire introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Locataires"
                ],
                "summary": "Mettre à jour un locataire (mise à jour partielle)",
                "operationId": "aacec60eb5a25b5918cb9184440f4a8c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du locataire",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "telephone": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Locataire mis à jour"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Locataire introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Locataire": {
                "title": "Locataire",
                "required": [
                    "nom",
                    "prenom"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Durand"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Alice"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "alice@example.com",
                        "nullable": true
                    },
                    "telephone": {
                        "type": "string",
                        "example": "07 98 76 54 32",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "title": "Erreur de validation",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "nom": [
                                "Le nom est obligatoire."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token Sanctum : Bearer {token}"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ],
...
}